Moved configurable widget dialog to configurable widget header
