+++ gui/include/industrial_calibration/gui/configurable_widget.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <QDialog>
 #include <QWidget>
 #include <yaml-cpp/yaml.h>
 
@@ -12,6 +13,13 @@ public:
 
     virtual void configure(const YAML::Node& node) = 0;
     virtual YAML::Node save() const = 0;
+};
+
+class ConfigurableWidgetDialog : public QDialog
+{
+public:
+    ConfigurableWidgetDialog(ConfigurableWidget* widget_, QWidget* parent = nullptr);
+    ConfigurableWidget* widget;
 };
 
 } // namespace industrial_calibration
